Написал тесты.

+++ src/components/list-page/logic-list-page.ts
@@ -34,18 +34,21 @@ export const addHeadList = async (parameters: IList) => {
   setValueInput('');
   setIsLoader({ ...isLoader, addHead: true });
   list.prepend(element);
-  listContainer[0].showAddCircle = true;
-  listContainer[0].valueSmall = valueInput;
-  listContainer[0].colorSmall = ElementStates.Changing;
-  setListContainer([...listContainer]);
+  let listArray = listContainer;
+  listArray[0].showAddCircle = true;
+  listArray[0].valueSmall = valueInput;
+  listArray[0].colorSmall = ElementStates.Changing;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[0].showAddCircle = false;
+  listArray = listContainer;
+  listArray[0].showAddCircle = false;
   listContainer.unshift(element);
-  listContainer[0].color = ElementStates.Modified;
-  setListContainer([...listContainer]);
+  listArray[0].color = ElementStates.Modified;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[0].color = ElementStates.Default;
-  setListContainer([...listContainer]);
+  listArray = listContainer;
+  listArray[0].color = ElementStates.Default;
+  setListContainer([...listArray]);
   setIsLoader({ ...isLoader, addHead: false });
 };
 
@@ -67,17 +70,20 @@ export const addTailList = async (parameters: IList) => {
   setValueInput('');
   setIsLoader({ ...isLoader, addTail: true });
   list.append(element);
-  listContainer[listLength - 1].showAddCircle = true;
-  listContainer[listLength - 1].valueSmall = valueInput;
-  listContainer[listLength - 1].colorSmall = ElementStates.Changing;
-  setListContainer([...listContainer]);
+  let listArray = listContainer;
+  listArray[listLength - 1].showAddCircle = true;
+  listArray[listLength - 1].valueSmall = valueInput;
+  listArray[listLength - 1].colorSmall = ElementStates.Changing;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
   listContainer.push(element);
-  listContainer[listLength - 1].showAddCircle = false;
-  listContainer[listLength].color = ElementStates.Modified;
-  setListContainer([...listContainer]);
+  listArray = listContainer;
+  listArray[listLength - 1].showAddCircle = false;
+  listArray[listLength].color = ElementStates.Modified;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[listLength].color = ElementStates.Default;
+  listArray = listContainer;
+  listArray[listLength].color = ElementStates.Default;
   setIsLoader({ ...isLoader, addTail: false });
 };
 
@@ -99,13 +105,15 @@ export const deleteElementList = async (parameters: IList, number: number) => {
   } else {
     setIsLoader({ ...isLoader, deleteTail: true });
   }
-  listContainer[number].showDeleteCircle = true;
-  listContainer[number].valueSmall = listContainer[number].value;
-  listContainer[number].value = '';
-  listContainer[number].colorSmall = ElementStates.Changing;
-  setListContainer([...listContainer]);
+  let listArray = listContainer;
+  listArray[number].showDeleteCircle = true;
+  listArray[number].valueSmall = listContainer[number].value;
+  listArray[number].value = '';
+  listArray[number].colorSmall = ElementStates.Changing;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[number].showDeleteCircle = false;
+  listArray = listContainer;
+  listArray[number].showDeleteCircle = false;
   if (number === 0) {
     listContainer.shift();
     list.deleteHead();
@@ -113,7 +121,7 @@ export const deleteElementList = async (parameters: IList, number: number) => {
     listContainer.pop();
     list.deleteTail();
   }
-  setListContainer([...listContainer]);
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
   if (number === 0) {
     setIsLoader({ ...isLoader, deleteHead: false });
@@ -149,30 +157,34 @@ export const addByIndexList = async (parameters: IList) => {
   list.addByIndex(element, Number(valueIndex));
 
   for (let i = 0; i < Number(valueIndex); i++) {
-    listContainer[i].showAddCircle = true;
-    listContainer[i].valueSmall = valueInput;
-    listContainer[i].colorSmall = ElementStates.Changing;
-    setListContainer([...listContainer]);
+    let listArray = listContainer;
+    listArray[i].showAddCircle = true;
+    listArray[i].valueSmall = valueInput;
+    listArray[i].colorSmall = ElementStates.Changing;
+    setListContainer([...listArray]);
     await delay(SHORT_DELAY_IN_MS);
-    listContainer[i].showAddCircle = false;
-    listContainer[i].color = ElementStates.Changing;
+    listArray = listContainer;
+    listArray[i].showAddCircle = false;
+    listArray[i].color = ElementStates.Changing;
   }
-  listContainer[Number(valueIndex)].showAddCircle = true;
-  listContainer[Number(valueIndex)].valueSmall = valueInput;
-  listContainer[Number(valueIndex)].colorSmall = ElementStates.Changing;
-  setListContainer([...listContainer]);
+  let listArray = listContainer;
+  listArray[Number(valueIndex)].showAddCircle = true;
+  listArray[Number(valueIndex)].valueSmall = valueInput;
+  listArray[Number(valueIndex)].colorSmall = ElementStates.Changing;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[Number(valueIndex)].showAddCircle = false;
-
-  listContainer.splice(Number(valueIndex), 0, element);
+  listArray = listContainer;
+  listArray[Number(valueIndex)].showAddCircle = false;
+  listArray.splice(Number(valueIndex), 0, element);
   for (let i = 0; i < Number(valueIndex); i++) {
-    listContainer[i].color = ElementStates.Default;
+    listArray[i].color = ElementStates.Default;
   }
-  listContainer[Number(valueIndex)].color = ElementStates.Modified;
-  setListContainer([...listContainer]);
+  listArray[Number(valueIndex)].color = ElementStates.Modified;
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-  listContainer[Number(valueIndex)].color = ElementStates.Default;
-  setListContainer([...listContainer]);
+  listArray = listContainer;
+  listArray[Number(valueIndex)].color = ElementStates.Default;
+  setListContainer([...listArray]);
   setIsLoader({ ...isLoader, addByIndex: false });
 };
 
@@ -193,30 +205,32 @@ export const deleteByIndexList = async (parameters: IList) => {
 
   setValueIndex('');
   setIsLoader({ ...isLoader, deleteByIndex: true });
+  if (Number(valueIndex) >= listLength) return;
 
   list.deleteByIndex(Number(valueIndex));
 
   for (let i = 0; i < Number(valueIndex); i++) {
-    listContainer[i].color = ElementStates.Changing;
-    setListContainer([...listContainer]);
+    let listArray = listContainer;
+    listArray[i].color = ElementStates.Changing;
+    setListContainer([...listArray]);
     await delay(SHORT_DELAY_IN_MS);
   }
   setListContainer([...listContainer]);
   await delay(SHORT_DELAY_IN_MS);
-
-  listContainer[Number(valueIndex)].showDeleteCircle = true;
-  listContainer[Number(valueIndex)].colorSmall = ElementStates.Changing;
-  listContainer[Number(valueIndex)].valueSmall =
-    listContainer[Number(valueIndex)].value;
-  listContainer[Number(valueIndex)].value = '';
-  setListContainer([...listContainer]);
+  let listArray = listContainer;
+  listArray[Number(valueIndex)].showDeleteCircle = true;
+  listArray[Number(valueIndex)].colorSmall = ElementStates.Changing;
+  listArray[Number(valueIndex)].valueSmall =
+    listArray[Number(valueIndex)].value;
+  listArray[Number(valueIndex)].value = '';
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
-
-  listContainer.splice(Number(valueIndex), 1);
+  listArray = listContainer;
+  listArray.splice(Number(valueIndex), 1);
   for (let i = 0; i < Number(valueIndex); i++) {
-    listContainer[i].color = ElementStates.Default;
+    listArray[i].color = ElementStates.Default;
   }
-  setListContainer([...listContainer]);
+  setListContainer([...listArray]);
   await delay(SHORT_DELAY_IN_MS);
   setIsLoader({ ...isLoader, deleteByIndex: false });
 };
